add asserts for Null

+++ includes/algorithms/matrix.h
@@ -19,11 +19,11 @@ void freeMemMatrices(matrix *ms, int nMatrices);
 
 void inputMatrix(matrix m);
 
-void inputMatrices(matrix *ms, size_t nMatrices);
+void inputMatrices(matrix *ms, int nMatrices);
 
 void outputMatrix(matrix m);
 
-void outputMatrices(matrix *ms, size_t nMatrices);
+void outputMatrices(matrix *ms, int nMatrices);
 
 void swapRows(matrix m, int i1, int i2);
